requirements.txt
ultralytics==8.0.196
python-dotenv
roboflow